Setup new embedding for plugin java code
